Update cfe.c
Fixed a significant bug in the  Eqn. that mimics Eqn. 2a from Jayawardena & Zhou (2000);  Added some shorter variable names to use notation consistent with that used in Jayawardena & Zhou (2000).  Also fixed the calculation of moisture content ratio Wm/Wmm which was incorrectly calculated using the ratio of  tension water storage/max tension water storage, when it correctly depends on the ratio of total water storage/max water storage.   This bug fix and code change are going to make a large difference in how and how well this code works when using Xinanjiang partitioning.